Prevent rate limit errors with curl.
While downloading kustomize with curl, a github rate limit can be hit
because this is an anonymous request from any github actions runner,
which is also used by other projects.

By using the ``GITHUB_TOKEN`` provided by github actions, we can let
the curl request be authenticated and make the request count towards
the rate limit of this generated token.